+++ src/atv_couch_wake/__init__.py
@@ -1,3 +1,3 @@
 """atv-couch-wake."""
 
-__version__ = "0.3.0"
+__version__ = "0.4.0"